+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Types of changes:
 * `Flows::SharedContextPipeline` wrap DSL, [issue](https://github.com/ffloyd/flows/issues/7)
 * `Flows::Flow` routing integrity check on initialization
 * `Flows::Plugin::OutputContract` skip contract DSL method
-* `Flows::Plugin::Profiler` introduced.
+* `Flows::Plugin::Profiler` introduced. Report types: raw, tree and flat.
 
 ### Changed